Add function playRound
The function takes two parameters: the option the player has chosen and
the option the computer has chosen. To make sure the text formats are
comparable, the humanChoice parameter is converted to lowercase letters.

If the computer and player chose the same option, the round results in
a draw. Otherwise, a comparison follows, based on the rules of the game:
Rock beats scissors, Scissors beat paper, paper beats rock. The score
variables for the computer and player are incremented accordingly.

No value is returned.